为 torch fsdp 添加断点重训；改善了关于 driver checkpoint 的测试

+++ tests/helpers/datasets/paddle_data.py
@@ -33,7 +33,7 @@ class PaddleNormalXYDataset(Dataset):
     def __getitem__(self, item):
         return {
             'x': paddle.to_tensor([self._data[item]], dtype='float32'),
-            'y': paddle.to_tensor([self._data[item]], dtype='float32')
+            'y': paddle.to_tensor([self._data[item]], dtype='int64')
         }
